server/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.11
pytest==7.4.0
